Refactor build system to support CSS modules (#237)
Reactist now generates a build more aligned to antd's best practices. It generates a clean ES6 build, a CommonJS build as well as an unpkg build. It also builds in support for future CSS module integration. This is a breaking change because importing modules has changed slightly. See README.md.

Co-authored-by: Vishnu Gopal <vishnugopal@users.noreply.github.com>